1. Adding products to wishlist table

+++ src/main/java/com/sagarannaldas/online_store/repositories/CategoryRepository.java
@@ -3,5 +3,5 @@ package com.sagarannaldas.online_store.repositories;
 import com.sagarannaldas.online_store.entities.Category;
 import org.springframework.data.repository.CrudRepository;
 
-public interface CategoryRepository extends CrudRepository<Category, Long> {
+public interface CategoryRepository extends CrudRepository<Category, Byte> {
 }
